spring mvc commit 4

+++ SPRING/springmvcs/src/main/java/com/bcits/springmvcs/bean/EmployeeInfoBean.java
@@ -1,6 +1,7 @@
 package com.bcits.springmvcs.bean;
 
 import java.io.Serializable;
+
 import java.util.Date;
 
 import javax.persistence.Column;
@@ -15,20 +16,15 @@ import lombok.Data;
 @Table(name="employee_primary_info")
 public class EmployeeInfoBean implements Serializable{
 	  
-//	  private int empId;
-//	  private String name;
-//	  private long mobileNo;
-//	  private String officialMailid;
-//	  private Date dateOfJoining;
-//	  private Double salary;
+
 	  
 	    @Id
 		@Column(name="emp_id")
-		private int empId;
+		private Integer empId;
 		@Column
 		private String name;
 		@Column(name="mobile_no")
-		private long mobileNo;
+		private Long mobileNo;
 		@Column(name="official_mail")
 		private String officialMailid;
 		@Column(name="dob")
@@ -40,13 +36,13 @@ public class EmployeeInfoBean implements Serializable{
 		@Column(name="blood_group")
 		private String bloodGroup;
 		@Column
-		private int salary;
+		private Double salary;
 		@Column(name="dept_id")
-		private int deptId;
+		private Integer deptId;
 		@Column(name="manager_id")
-		private int managerId;
+		private Integer managerId;
 		@Column
 		private String password;
-
+        
 		
 }
